test: [SDK-5048] split the feature-flag tests by unit under test
The single file had grown past the 400-line SwiftLint limit as the refresh
service picked up lifecycle coverage. Moves the refresh-service cases and their
fakes into their own file, leaving the manager and backend cases behind, so both
sit well under the limit and each file covers one unit. `StubFeatureFlagsHttp`
is shared by both and becomes internal.
